[Feature/comment] JoI 유효성 검사 추가

+++ src/controllers/comment.controller.js
@@ -1,3 +1,4 @@
+const { BadRequestError } = require('../exceptions/customError.js');
 const CommentService = require('../services/comment.service');
 const { logger } = require('../middlewares/logger');
 
@@ -5,11 +6,18 @@ class CommentController {
   constructor() {
     this.commentService = new CommentService();
   }
-  getAllComment = async (req, res) => {
+  getAllComment = async (req, res, next) => {
     logger.info(`CommentController.getAllComment Request`);
     const { pinId } = req.params;
-    const comments = await this.commentService.getAllComment({ pinId });
-    res.status(200).json({ comments });
+    if (!pinId) {
+      throw new BadRequestError('파라미터 값이 올바르지 않습니다.');
+    }
+    try {
+      const comments = await this.commentService.getAllComment({ pinId });
+      res.status(200).json({ comments });
+    } catch (error) {
+      next(error);
+    }
   };
 
   addComment = async (req, res, next) => {
@@ -18,6 +26,13 @@ class CommentController {
     const userId = Math.floor(Math.random() * 11);
     const { content } = req.body;
     try {
+      if (!pinId) {
+        throw new BadRequestError('파라미터 값이 올바르지 않습니다.');
+      }
+      if (!content) {
+        throw new BadRequestError('내용을 입력해 주세요.');
+      }
+
       const comment = await this.commentService.addComment({
         userId,
         pinId,
@@ -34,6 +49,9 @@ class CommentController {
     logger.info(`CommentController.deleteComment Request`);
     const userId = Math.floor(Math.random() * 11);
     const { pinId, commentId } = req.params;
+    if (!pinId || !commentId) {
+      throw new BadRequestError('파라미터 값이 올바르지 않습니다.');
+    }
     try {
       const comment = await this.commentService.deleteComment({
         userId,
